+++ src/demo/utils/demos/vertical.vue
@@ -1,0 +1,14 @@
+<template>
+  <TvProgressBar
+    orientation="vertical"
+    position="left"
+    width="6px"
+    color="#4f46e5"
+    glow
+  />
+</template>
+
+<script setup>
+import { TvProgressBar } from "@todovue/tv-progress-bar";
+import "@todovue/tv-progress-bar/style.css";
+</script>